feat: rewrite 1ES compiler to use Copilot CLI + AWF + MCPG
Replace the legacy Agency job type (agencyJob) with direct Copilot CLI
invocation, AWF network isolation, and MCP Gateway — matching the
standalone pipeline execution model.

Changes:
- templates/1es-base.yml: Complete rewrite. All three jobs
  (PerformAgenticTask, AnalyzeSafeOutputs, ProcessSafeOutputs) now use
  templateContext.type: buildJob with the same step sequence as
  standalone. Dropped Agency concepts: commandOptions, globalOptions,
  logLevel, mcpConfiguration, agentContextRoot, AgencyArtifact.
- src/compile/onees.rs: Rewritten as thin wrapper using compile_shared().
  Removed generate_agent_context_root, generate_mcp_configuration,
  generate_inline_steps. Only 1ES-specific setup/teardown helpers remain.
- src/compile/common.rs: Moved generate_allowed_domains from standalone.
  Removed dead is_custom_mcp function.
- src/compile/standalone.rs: Removed now-unnecessary imports.

Both compilers now share the same execution model and compile flow.
Standalone output is verified byte-identical to pre-refactor baseline.
